% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut-package.R
\docType{package}
\name{ut-package}
\alias{ut}
\alias{ut-package}
\title{ut: UKHSA Tools/Utility Tools}
\description{
Various utility functions for analysis performed on the UKHSA/EY Project on Evaluating COVID-19 Testing Strategies in the UK undertaken in late 2022.
}
\keyword{internal}
